Basic Scene
Spawns a grid
You can drag tiles around and they will swap with their neighbours

3D Mouse
Bit off. When you drag tiles, they change their height based on the mouse.
So if you drag a tile from the bottom of the screen to the top, it will be
way above and cast a shadow. Dragging from the top tiles to the bottom,
the tile will be underneath.
Still works
Once you let go of the mouse, they go where they're supposed to.