Building Towers
Can press a button and build a tower at the selection

Building Tower
Check if have a selection
Check if the tile can be built on (just checks if it already has a tower)
Build a Tower. Handled by TowerStore. It instantiates and changes the
parent and tells the grid that it built on selection
Grid tells the tile it now has a tower on it.

Kinda fragile and sloppy
Checking if there is a current target.
Instead of a bool and a target, should be an Option<target>

Selection is Awkward
You click to select and click to deselect and click to buy a tower in the menu
This should probably be in one place. When you click, it figures out what
to do and then sends an event with the correct thing.
Right now, you click on tiles to select them and click on the deselect
plane to select nothing (basically a huge X button) and click on the
button to build. The button can't overlap the deselect plane or they both
are clicked.

Fixed Bug
When you selected a tile, built a tower, then moved the tower, then build
again, you would built a tower at the place where you last built, but it
wasn't selected anymore.
When swapping, the current selection in the grid wasn't updated. It was
only being updated when first clicked on
Fixed, but something similar could probably show up again.