Gold Spawner
Made a tower that spawns gold around it.
The gold sticks to the tile it is spawned on.
When you swap tiles, the gold from both tiles gets combined. Not quite
working 100%. Sometimes you can drag an empty tile and take the gold off
of a tile with gold on it, sometimes it doesn't work...

Resource
Just gold for now, but should be able to work for ore, wood, etc.

I'm not very happy with this implementation of it, but it works.