Gold Spawning Refactor
Tiles store the amount of gold they have
This is how I originally did it.
The version with spawning gold objects was too messy with keeping track of
the count of objects, where they were, and their parents.

Deleted Resource.cs

ResourceVisualizer.cs
Attached to the tile objects
When the tile gold count changes, this handles the visuals
It spawns a gold sphere for each gold in a stack.
When the count changes, it deletes all the gold and spawns new gold. Could
be better, but oh well.